Add custom keybindings for mini.bracketed to go to error messages.
mini.bracketed provides MiniBracketed.diagnostic() which jumps to
diagnostic sequentially. By default it jumps to any diagnostic, which we
sometimes don't want since there may be a lot of them and we are only
interested in fixing errors. The module doesn't provide a setting to
jump to diagnostics of different severities conditionally, so we have to
bind our own.