Fix: keymap change when enter non-modifable buffers.
The previous commit changed so that whenever we enter a buffer, set the
keymap to "vietnamese-telex_utf-8" and then turn it off because we want
English by default. However, the implementation require that we enter
insert mode to turn off the keymap, which is imposible in unmodifiable
buffers and then neovim beeps at us.

This commit fix that by check if the buffer is modifiable or not before
doing any of that stuff.